+++ tests/tests/cloudinary_/models/test_image.py
@@ -73,6 +73,10 @@ class TestCloudinaryImage(CloudinaryFileResource):
             'id': 1,
             'name': self.resource_name,
             'extension': self.resource_extension,
+            'caption': '{}.{}'.format(
+                self.resource_name,
+                self.resource_extension
+            ),
             'size': self.resource_size,
             'width': 1534,
             'height': 2301,