+++ tests/tests/cloudinary_/models/test_media.py
@@ -75,6 +75,10 @@ class TestCloudinaryMedia(CloudinaryFileResource):
             'id': 1,
             'name': self.resource_name,
             'extension': self.resource_extension,
+            'caption': '{}.{}'.format(
+                self.resource_name,
+                self.resource_extension
+            ),
             'size': self.resource_size,
             'file_info': '(mp3, 2.1\xa0MB)',
             'url': storage.resource.get_file_url(),