+++ tests/tests/models/test_collection.py
@@ -400,8 +400,11 @@ class TestFileItem(CollectionItemMixin, TestFileFieldResource):
             'itemType': 'file',
             'name': self.resource_name,
             'extension': self.resource_extension,
+            'caption': '{}.{}'.format(
+                self.resource_name,
+                self.resource_extension
+            ),
             'size': self.resource_size,
-            'caption': '{}.{}'.format(self.resource_name, self.resource_extension),
             'order': 0,
             'preview': render_to_string(
                 'paper_uploads/items/preview/file.html',
@@ -598,8 +601,11 @@ class TestSVGItem(CollectionItemMixin, TestFileFieldResource):
             'itemType': 'svg',
             'name': self.resource_name,
             'extension': self.resource_extension,
+            'caption': '{}.{}'.format(
+                self.resource_name,
+                self.resource_extension
+            ),
             'size': self.resource_size,
-            'caption': '{}.{}'.format(self.resource_name, self.resource_extension),
             'order': 0,
             'preview': render_to_string(
                 'paper_uploads/items/preview/svg.html',
@@ -800,13 +806,16 @@ class TestImageItem(CollectionItemMixin, TestFileFieldResource):
             'itemType': 'image',
             'name': self.resource_name,
             'extension': self.resource_extension,
+            'caption': '{}.{}'.format(
+                self.resource_name,
+                self.resource_extension
+            ),
             'size': self.resource_size,
             'width': 1534,
             'height': 2301,
             'cropregion': '',
             'title': '',
             'description': '',
-            'caption': '{}.{}'.format(self.resource_name, self.resource_extension),
             'order': 0,
             'preview': render_to_string(
                 'paper_uploads/items/preview/image.html',