+++ tests/tests/models/test_file.py
@@ -53,6 +53,10 @@ class TestUploadedFile(TestFileFieldResource):
             'id': 1,
             'name': self.resource_name,
             'extension': self.resource_extension,
+            'caption': '{}.{}'.format(
+                self.resource_name,
+                self.resource_extension
+            ),
             'size': self.resource_size,
             'file_info': '(Jpeg, 672.8\xa0KB)',
             'url': storage.resource.get_file_url(),